@@ -1,7 +1,7 @@
 {
   "name": "@firestitch/form",
   "title": "Form",
-  "version": "7.6.8",
+  "version": "8.0.0",
   "repository": {
     "type": "git",
     "url": "https://github.com/Firestitch/ngx-form"
@@ -27,60 +27,62 @@
     "serve": "cd tools && npm run cli:serve"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.12.2",
-    "@angular-devkit/build-ng-packagr": "^0.12.2",
-    "@angular/animations": "^7.2.1",
-    "@angular/cdk": "^7.3.7",
-    "@angular/cli": "^7.2.2",
-    "@angular/common": "^7.2.1",
-    "@angular/compiler": "^7.2.1",
-    "@angular/compiler-cli": "^7.2.1",
-    "@angular/core": "^7.2.1",
-    "@angular/flex-layout": "^7.0.0-beta.23",
-    "@angular/forms": "^7.2.1",
-    "@angular/http": "^7.2.1",
-    "@angular/material": "^7.3.7",
-    "@angular/platform-browser": "^7.2.1",
-    "@angular/platform-browser-dynamic": "^7.2.1",
-    "@angular/router": "^7.2.10",
-    "@compodoc/compodoc": "^1.1.7",
-    "@firestitch/autocomplete-chips": "^7.1.8",
-    "@firestitch/checkboxgroup": "^7.1.1",
+    "@angular-devkit/build-angular": "^0.803.9",
+    "@angular-devkit/build-ng-packagr": "^0.803.9",
+    "@angular/animations": "^8.2.10",
+    "@angular/cdk": "^8.2.3",
+    "@angular/cli": "^8.3.9",
+    "@angular/common": "^8.2.10",
+    "@angular/compiler": "^8.2.10",
+    "@angular/compiler-cli": "^8.2.10",
+    "@angular/core": "^8.2.10",
+    "@angular/flex-layout": "^8.0.0-beta.27",
+    "@angular/forms": "^8.2.10",
+    "@angular/http": "^7.2.15",
+    "@angular/material": "^8.2.3",
+    "@angular/platform-browser": "^8.2.10",
+    "@angular/platform-browser-dynamic": "^8.2.10",
+    "@angular/router": "^8.2.10",
+    "@compodoc/compodoc": "^1.1.10",
+    "@firestitch/api": "^8.0.1",
+    "@firestitch/autocomplete-chips": "^8.0.0",
+    "@firestitch/checkboxgroup": "^7.1.10",
     "@firestitch/clear": "^7.1.0",
     "@firestitch/common": "^7.0.2",
-    "@firestitch/datepicker": "^7.1.1",
-    "@firestitch/example": "^7.0.0",
-    "@firestitch/label": "^7.1.2",
-    "@firestitch/message": "^7.1.4",
-    "@firestitch/phone": "^7.1.0",
-    "@firestitch/radiogroup": "^7.2.3",
-    "@firestitch/skeleton": "^7.1.0",
-    "@firestitch/password": "^7.0.3",
-    "@firestitch/api": "^7.0.2",
-    "codelyzer": "^4.5.0",
+    "@firestitch/datepicker": "^8.0.1",
+    "@firestitch/drawer": "^8.0.0",
+    "@firestitch/example": "^8.0.0",
+    "@firestitch/iframe": "^8.0.0",
+    "@firestitch/label": "^7.1.9",
+    "@firestitch/menu": "^7.0.11",
+    "@firestitch/message": "^8.0.1",
+    "@firestitch/password": "^7.0.4",
+    "@firestitch/phone": "^7.2.1",
+    "@firestitch/radiogroup": "^7.2.5",
+    "@firestitch/scroll-picker": "^7.0.3",
+    "@firestitch/skeleton": "^7.1.2",
+    "codelyzer": "^5.1.2",
     "cpx": "^1.5.0",
-    "date-fns": "^2.0.0-alpha.27",
+    "date-fns": "^2.4.1",
     "hammerjs": "^2.0.8",
+    "highlight.js": "^9.15.10",
+    "imask": "^5.2.1",
     "iscroll": "^5.2.0",
-    "lodash-es": "^4.17.0",
-    "ng-packagr": "^4.6.0",
+    "lodash-es": "^4.17.15",
+    "ng-packagr": "^5.6.1",
+    "ngx-toastr": "^11.1.0",
     "npm-run-all": "^4.1.5",
     "rxjs": "^6.5.3",
-    "tsickle": "^0.34.0",
-    "tslint": "^5.12.1",
-    "typescript": "^3.1.6",
-    "webpack": "^4.29.0",
-    "webpack-dev-server": "^3.1.14",
-    "zone.js": "^0.8.28"
+    "tsickle": "^0.37.0",
+    "tslint": "^5.20.0",
+    "typescript": "~3.5.3",
+    "zone.js": "^0.10.2"
   },
   "engines": {
     "node": ">=6.0.0"
   },
   "peerDependencies": {
     "rxjs": ">=6.3.0",
-    "zone.js": ">=0.8.4",
-    "@firestitch/common": ">=7.0.2",
-    "date-fns": ">=2.0.0-alpha.2",
-    "lodash-es": ">=4.17.0"
+    "@firestitch/common": ">=7.0.2"
   }
 }
